Refact: Implemented CreateTeamEditor
