+++ src/app/recruit/write/panel/fields/SelectRegion.tsx
@@ -2,7 +2,7 @@ import MenuItem from '@mui/material/MenuItem'
 import Select from '@mui/material/Select'
 import { koreaDistrict } from '@/constant/DistrictData'
 import { Stack, Typography } from '@mui/material'
-import * as style from './SelectRegion.style'
+import * as style from '../../../[id]/edit/panel/SelectRegion.style'
 import { Control, Controller, useFormState } from 'react-hook-form'
 import { IRecruitWriteField } from '@/app/recruit/write/page'
